# docs/next-phase/openapi-v1.json
{
  "components": {
    "schemas": {
      "AbortRequest": {
        "additionalProperties": false,
        "properties": {
          "reason": {
            "maxLength": 2000,
            "minLength": 3,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "AbortRequest",
        "type": "object"
      },
...
      "ApprovalRequest": {
        "additionalProperties": false,
        "properties": {
          "operator_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operator Id"
          },
          "reason": {
            "maxLength": 2000,
            "minLength": 3,
            "title": "Reason",
            "type": "string"
          },
          "revision_digest": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Revision Digest",
            "type": "string"
          }
        },
        "required": [
          "revision_digest",
          "reason"
        ],
        "title": "ApprovalRequest",
        "type": "object"
      },
...
      "CommitRequest": {
        "additionalProperties": false,
        "properties": {
          "revision_digest": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Revision Digest",
            "type": "string"
          },
          "step_delay_ms": {
            "default": 0,
            "description": "demo pacing only",
            "maximum": 3000.0,
            "minimum": 0.0,
            "title": "Step Delay Ms",
            "type": "integer"
          }
        },
        "required": [
          "revision_digest"
        ],
        "title": "CommitRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "PACT - Protocol for Agentic Commit Transactions",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/transactions": {
      "get": {
        "operationId": "list_transactions_api_v1_transactions_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 500,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 40,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "in": "query",
            "name": "workflow",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workflow"
            }
          },
          {
            "in": "query",
            "name": "actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Actor"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "created_after",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created After"
            }
          },
          {
            "in": "query",
            "name": "created_before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created Before"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Transactions Api V1 Transactions Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Transactions",
        "tags": [
          "transactions"
        ]
      },
      "post": {
        "operationId": "create_transaction_api_v1_transactions_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-PACT-Request-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Pact-Request-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BeginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Transaction Api V1 Transactions Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Transaction",
        "tags": [
          "transactions"
        ]
      }
    },
    "/api/v1/transactions/queues/approvals": {
      "get": {
        "operationId": "approval_queue_api_v1_transactions_queues_approvals_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Approval Queue Api V1 Transactions Queues Approvals Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approval Queue",
        "tags": [
          "transactions"
        ]
      }
    },
    "/api/v1/transactions/queues/incidents": {
      "get": {
        "operationId": "incident_queue_api_v1_transactions_queues_incidents_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Incident Queue Api V1 Transactions Queues Incidents Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Incident Queue",
        "tags": [
          "transactions"
        ]
      }
    },
...
    "/api/v1/transactions/{tx_id}": {
      "get": {
        "operationId": "get_transaction_api_v1_transactions__tx_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "tx_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tx Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Transaction Api V1 Transactions  Tx Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Transaction",
        "tags": [
          "transactions"
        ]
      }
    },
    "/api/v1/transactions/{tx_id}/abort": {
      "post": {
        "operationId": "abort_api_v1_transactions__tx_id__abort_post",
        "parameters": [
          {
            "in": "path",
            "name": "tx_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tx Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AbortRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Abort Api V1 Transactions  Tx Id  Abort Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Abort",
        "tags": [
          "transactions"
        ]
      }
    },
    "/api/v1/transactions/{tx_id}/approve": {
      "post": {
        "operationId": "approve_api_v1_transactions__tx_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "tx_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tx Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Approve Api V1 Transactions  Tx Id  Approve Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve",
        "tags": [
          "transactions"
        ]
      }
    },
    "/api/v1/transactions/{tx_id}/children": {
      "post": {
        "operationId": "create_child_api_v1_transactions__tx_id__children_post",
        "parameters": [
          {
            "in": "path",
            "name": "tx_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tx Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-PACT-Request-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Pact-Request-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DelegateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Child Api V1 Transactions  Tx Id  Children Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Child",
        "tags": [
          "transactions"
        ]
      }
    },
    "/api/v1/transactions/{tx_id}/commit": {
      "post": {
        "operationId": "commit_api_v1_transactions__tx_id__commit_post",
        "parameters": [
          {
            "in": "path",
            "name": "tx_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tx Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Commit Api V1 Transactions  Tx Id  Commit Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Commit",
        "tags": [
          "transactions"
        ]
      }
    },
    "/api/v1/transactions/{tx_id}/commit-decision": {
      "get": {
        "operationId": "commit_decision_api_v1_transactions__tx_id__commit_decision_get",
        "parameters": [
          {
            "in": "path",
            "name": "tx_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tx Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Commit Decision Api V1 Transactions  Tx Id  Commit Decision Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Commit Decision",
        "tags": [
          "transactions"
        ]
      }
    },
    "/api/v1/transactions/{tx_id}/effects": {
      "post": {
        "operationId": "propose_effect_api_v1_transactions__tx_id__effects_post",
        "parameters": [
          {
            "in": "path",
            "name": "tx_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tx Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-PACT-Request-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Pact-Request-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EffectProposal"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Propose Effect Api V1 Transactions  Tx Id  Effects Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Propose Effect",
        "tags": [
          "transactions"
        ]
      }
    },
    "/api/v1/transactions/{tx_id}/effects/{effect_id}/withdraw": {
      "post": {
        "operationId": "withdraw_api_v1_transactions__tx_id__effects__effect_id__withdraw_post",
        "parameters": [
          {
            "in": "path",
            "name": "tx_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tx Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "effect_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Effect Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WithdrawRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Withdraw Api V1 Transactions  Tx Id  Effects  Effect Id  Withdraw Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Withdraw",
        "tags": [
          "transactions"
        ]
      }
    },
...
    "/api/v1/transactions/{tx_id}/operator-actions": {
      "post": {
        "operationId": "operator_action_api_v1_transactions__tx_id__operator_actions_post",
        "parameters": [
          {
            "in": "path",
            "name": "tx_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tx Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OperatorActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Operator Action Api V1 Transactions  Tx Id  Operator Actions Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Operator Action",
        "tags": [
          "transactions"
        ]
      }
    },
    "/api/v1/transactions/{tx_id}/prepare": {
      "post": {
        "operationId": "prepare_api_v1_transactions__tx_id__prepare_post",
        "parameters": [
          {
            "in": "path",
            "name": "tx_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tx Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Prepare Api V1 Transactions  Tx Id  Prepare Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Prepare",
        "tags": [
          "transactions"
        ]
      }
    },
    "/api/v1/transactions/{tx_id}/projection": {
      "get": {
        "operationId": "projection_api_v1_transactions__tx_id__projection_get",
        "parameters": [
          {
            "in": "path",
            "name": "tx_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tx Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Projection Api V1 Transactions  Tx Id  Projection Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Projection",
        "tags": [
          "transactions"
        ]
      }
    },
...
  }
}